Patch 0165
・高速化のため map_charid2sd() を手抜きマクロから関数に変更（map.*）
・map_quit() を少し整理（map.c）
・map_freeblock_timer() を呼び出す間隔を1分から10分にしてみる（map.c）
	エラー自体見たことないのでこの処理を丸ごと削除しても良さそうだが・・・
・pc.c, pc.hの整理
・motdの表示処理を少しだけ改善（pc.c）
・skill_tree構造体のneedメンバの配列が一つ多かったので修正（pc.c）

+++ src/map/skill.c
@@ -1264,7 +1264,7 @@ int skill_additional_effect( struct block_list* src, struct block_list *bl,int s
 			{
 				if(atn_rand()%10000 < sd->break_myequip_rate_when_attack[i])
 				{
-					pc_break_equip2(sd,(unsigned short)i);
+					pc_break_equip2(sd,i);
 				}
 			}
 		}
